Replace "except.hpp" with "exceptions.hpp" in multiple files because of file renaming.
