+++ bevarmejolib/include/bevarmejo/evaluation/metric.hpp
@@ -6,7 +6,7 @@
 #include <tuple>
 #include <type_traits>
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 #include "bevarmejo/utility/quantity_series.hpp"
 #include "bevarmejo/utility/type_traits.hpp"
 #include "bevarmejo/wds/water_distribution_system.hpp"
