+++ bevarmejolib/include/bevarmejo/utility/pagmo/serializers/json/default_objects.hpp
@@ -15,7 +15,7 @@
 #include "bevarmejo/io/json.hpp"
 #include "bevarmejo/io/aliased_key.hpp"
 #include "bevarmejo/io/keys/bemeopt.hpp"
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 #include "bevarmejo/utility/string.hpp"
 
 namespace bevarmejo::io::key::detail
