+++ bevarmejolib/include/bevarmejo/utility/registry.hpp
@@ -5,7 +5,7 @@
 #include <utility>
 #include <vector>
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 #include "bevarmejo/utility/memory.hpp"
 
 #include "bevarmejo/utility/registry_users.ipp"
