+++ bevarmejolib/include/bevarmejo/utility/registry_view.hpp
@@ -7,7 +7,7 @@
 #include <type_traits>
 #include <vector>
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 #include "bevarmejo/utility/memory.hpp"
 #include "bevarmejo/utility/registry.hpp"
 #include "bevarmejo/utility/unique_string_sequence.hpp"
