+++ bevarmejolib/include/bevarmejo/wds/water_distribution_system.hpp
@@ -24,7 +24,7 @@ namespace fsys = std::filesystem;
 #include "bevarmejo/wds/utility/epanet/en_help.hpp"
 
 #include "bevarmejo/io/streams.hpp"
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 
 #include "bevarmejo/utility/registry.hpp"
 #include "bevarmejo/utility/registry_view.hpp"
