+++ bevarmejolib/src/evaluation/metrics/hydraulic_functions.cpp
@@ -1,5 +1,5 @@
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 #include "bevarmejo/wds/utility/quantity_series.hpp"
 #include "bevarmejo/wds/water_distribution_system.hpp"
