+++ bevarmejolib/src/simulation/hyd_sim_settings.cpp
@@ -1,4 +1,4 @@
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 #include "bevarmejo/utility/time.hpp"
 
 #include "bevarmejo/simulation/hyd_sim_settings.hpp"
