+++ bevarmejolib/src/wds/elements/curve.cpp
@@ -2,7 +2,7 @@
 
 #include "epanet2_2.h"
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 
 #include "bevarmejo/wds/utility/quantity_series.hpp"
 #include "bevarmejo/wds/element.hpp"
