+++ bevarmejolib/src/wds/elements/network_elements/nodes/reservoir.cpp
@@ -2,7 +2,7 @@
 
 #include "epanet2_2.h"
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 
 #include "bevarmejo/wds/element.hpp"
 #include "bevarmejo/wds/elements/network_element.hpp"
