+++ bevarmejolib/src/wds/elements/pattern.cpp
@@ -4,7 +4,7 @@
 
 #include "epanet2_2.h"
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 
 #include "bevarmejo/wds/element.hpp"
