+++ bevarmejolib/src/wds/utility/epanet/water_distribution_system.cpp
@@ -14,7 +14,7 @@ namespace fsys = std::filesystem;
 #include "epanet2_2.h"
 #include "types.h"
 
-#include "bevarmejo/utility/except.hpp"
+#include "bevarmejo/utility/exceptions.hpp"
 #include "bevarmejo/utility/metadata.hpp"
 
 #include "bevarmejo/io/streams.hpp"
